+++ supvisors/supvisorsctl.py
@@ -118,10 +118,12 @@ class ControllerPlugin(ControllerPluginBase):
             else:
                 # create template. identifier has variable length
                 max_identifiers = ControllerPlugin.max_template(info_list, 'identifier', 'Supervisor')
-                template = f'%(identifier)-{max_identifiers}s%(state)-11s%(load)-6s%(ltime)-10s%(counter)-9s'
+                max_node_names = ControllerPlugin.max_template(info_list, 'node_name', 'Node')
+                template = (f'%(identifier)-{max_identifiers}s%(node_name)-{max_node_names}s%(port)-7s%(state)-11s'
+                            '%(load)-6s%(ltime)-10s%(counter)-9s')
                 # print title
-                payload = {'identifier': 'Supervisor', 'state': 'State', 'load': 'Load',
-                           'ltime': 'Time', 'counter': 'Counter'}
+                payload = {'identifier': 'Supervisor', 'node_name': 'Node', 'port': 'Port', 'state': 'State',
+                           'load': 'Load', 'ltime': 'Time', 'counter': 'Counter'}
                 self.ctl.output(template % payload)
                 # check request args
                 identifiers = arg.split()
@@ -129,7 +131,9 @@ class ControllerPlugin(ControllerPluginBase):
                 # print filtered payloads
                 for info in info_list:
                     if output_all or info['identifier'] in identifiers:
-                        payload = {'identifier': info['identifier'], 'state': info['statename'],
+                        payload = {'identifier': info['identifier'],
+                                   'node_name': info['node_name'], 'port': info['port'],
+                                   'state': info['statename'],
                                    'load': '{}%'.format(info['loading']),
                                    'counter': info['sequence_counter'],
                                    'ltime': simple_localtime(info['local_time'])}