+++ supvisors/tests/test_supvisorsctl.py
@@ -253,9 +253,9 @@ def test_sstate(controller, plugin, mocked_check):
 def test_instance_status(controller, plugin, mocked_check):
     """ Test the instance_status request. """
     mocked_rpc = plugin.supvisors().get_all_instances_info
-    mocked_rpc.return_value = [{'identifier': '10.0.0.1',
+    mocked_rpc.return_value = [{'identifier': '10.0.0.1', 'node_name': '10.0.0.1', 'port': 60000,
                                 'statename': 'running', 'loading': 10, 'local_time': 1500, 'sequence_counter': 12},
-                               {'identifier': '10.0.0.2',
+                               {'identifier': '10.0.0.2', 'node_name': '10.0.0.2', 'port': 60000,
                                 'statename': 'stopped', 'loading': 0, 'local_time': 100, 'sequence_counter': 15}]
     _check_call(controller, mocked_check, mocked_rpc,  plugin.help_instance_status, plugin.do_instance_status,
                 '', [call()])